% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{[[.transit_map}
\alias{[[.transit_map}
\title{Look up a key in a transit map}
\usage{
\method{[[}{transit_map}(x, i, ...)
}
\arguments{
\item{x}{a `transit_map()`.}

\item{i}{the key.}

\item{...}{unused.}
}
\description{
`m[[key]]` finds the value for `key` (any transit value) in a
`transit_map()`, or `NULL` if it isn't there.
}
