% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{transit_collections}
\alias{transit_collections}
\alias{transit_map}
\alias{transit_set}
\alias{transit_list}
\alias{transit_tagged}
\alias{transit_link}
\title{Transit collections and tagged values}
\usage{
transit_map(keys = list(), values = list())

transit_set(elements = list())

transit_list(elements = list())

transit_tagged(tag, rep)

transit_link(href, rel, name = NULL, prompt = NULL, render = NULL)
}
\arguments{
\item{keys, values}{lists of the same length.}

\item{elements}{a list.}

\item{tag}{a string.}

\item{rep}{the tagged value's representation, any transit value.}

\item{href, rel, name, prompt, render}{the parts of a link; `href` is a
`transit_uri()` or a string.}
}
\description{
`transit_map()` is a map whose keys can be any transit value (R named
lists only allow strings); `transit_set()` a set; `transit_list()` a
transit list, as distinct from an array (which is a plain R list).
`transit_tagged()` is a value with a tag this package doesn't interpret,
which is written back out unchanged. `transit_link()` is a transit link.
}
