% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{transit_equal}
\alias{transit_equal}
\title{Compare transit values}
\usage{
transit_equal(a, b)
}
\arguments{
\item{a, b}{values to compare.}
}
\value{
`TRUE` or `FALSE`.
}
\description{
Like [identical()], except that maps and sets compare regardless of order
and `NaN` equals `NaN`.
}
