% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{transit_format}
\alias{transit_format}
\title{Format transit values}
\usage{
transit_format(x, ...)
}
\arguments{
\item{x}{a value.}

\item{...}{unused.}
}
\description{
Renders a value in an EDN-like notation, as used by `print()` for the
transit classes.
}
