% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{transit_types}
\alias{transit_types}
\alias{transit_keyword}
\alias{transit_symbol}
\alias{transit_uuid}
\alias{transit_uri}
\alias{transit_decimal}
\alias{transit_bigint}
\alias{transit_int64}
\title{Transit value constructors}
\usage{
transit_keyword(x)

transit_symbol(x)

transit_uuid(x)

transit_uri(x)

transit_decimal(x)

transit_bigint(x)

transit_int64(x)
}
\arguments{
\item{x}{a string (or for `transit_int64()`, `transit_bigint()` and
`transit_decimal()`, a number).}
}
\description{
Constructors for transit types that have no direct R equivalent. Most hold
a single string: `transit_int64()` a 64 bit integer outside R's integer
range, `transit_bigint()` an arbitrary precision integer and
`transit_decimal()` an arbitrary precision decimal, each in decimal form.
}
